Add Error Check for Editing Files
This check was found thanks to Amazon. It now handles the scenario
where the skill would search for a file to edit, and when one was
not found it would return something that Alexa could not handle.
Now, it returns an error message to say that file was not found.
Skill should be accepted for publishing.  Resubmitted now.